Added the final output file and the ISO 639-1 Data
apiresp.json is a JSON file that app_.js will write to after running the third party API, and theapp_cb.py will read from it to send the output to the user. langs.json contains the ISO 639-1 Data the abbreviates a language to a code; e.g. Arabic's ISO 639-1 is ar; e.g. English's ISO is eg and so on.